// app-turnOver-v1/src/app/rotas.enum.ts
export enum Rotas {
  Home = '',
  Login = 'login',
  Dashboard = 'paginas/dashboard/grafico-anual',
  IaIndica = 'paginas/ia-indica',

  // Cadastros
  CadastroFuncionario = 'paginas/paginas-cadastro/cadastro-funcionario',
  CadastroAcompanhamento = 'paginas/paginas-cadastro/cadastro-acompanhamento',
  CadastroFitCultural = 'paginas/paginas-cadastro/cadastro-fit-cultural',
  CadastroDesligamento = 'paginas/paginas-cadastro/cadastro-desligamento',
  CadastroSetor = 'paginas/paginas-cadastro/cadastro-setor',
  CadastroCargo = 'paginas/paginas-cadastro/cadastro-cargo',
  CadastroMotivoDesligamento = 'paginas/paginas-cadastro/cadastro-motivo-desligamento',
  // Listagem
  ListagemFuncionarios = 'paginas/paginas-listagem/listagem-funcionarios',
  ListagemSetores = 'paginas/paginas-listagem/listagem-setores',
  ListagemCargos = 'paginas/paginas-listagem/listagem-cargos',
  ListagemFitCultural = 'paginas/paginas-listagem/listagem-fit-cultural',
  ListagemDesligamentos = 'paginas/paginas-listagem/listagem-desligamentos',
  ListagemAcompanhamentos = 'paginas/paginas-listagem/listagem-acompanhamentos',
  ListagemMotivoDesligamento = 'paginas/paginas-listagem/listagem-motivo-desligamento',
}